Added NSSet and NSDictionary
In order to prevent a course from having the same student objects
multiple times, and from having student objects have the same course
object multiple times, I made the collections that handle that
information an NSSet, so that the object can only be added once.
An NSDictionary now handles a ‘bio’ for teachers. Theoretically, an
NSDictionary could have stored all Person object info. But it was too
late to switch things over.

BUG:
When associating a teacher to a course, if you were to associate
another teacher to the same course, the course will correctly reflect
the new teacher change. However, the course still appears under the old
teachers ‘associated courses’. Will need to make a method that removes
an associated course from the Staff object.